Added scripts to check azure blob content
